Clean and rename example components
