+++ src/components/atoms/ExampleBox/ExampleBox.module.scss
