+++ src/components/atoms/ExampleButton/ExampleButton.stories.tsx
@@ -3,17 +3,17 @@ import React from 'react';
 import { Story, Meta } from '@storybook/react/types-6-0';
 // import { Story, Meta } from '@storybook/react';
 
-import { Button, ButtonProps } from './Button';
+import { ExampleButton, ExampleButtonProps } from './ExampleButton';
 
 export default {
   title: 'Example/Button',
-  component: Button,
+  component: ExampleButton,
   argTypes: {
     backgroundColor: { control: 'color' },
   },
 } as Meta;
 
-const Template: Story<ButtonProps> = (args) => <Button {...args} />;
+const Template: Story<ExampleButtonProps> = (args) => <ExampleButton {...args} />;
 
 export const Primary = Template.bind({});
 Primary.args = {
